Drop unused prepare hook; move localInfoPrefix out of hover.ts
prepare never actually fires: CI/packaging always go through root's
gen -> compile/webpack chain (vsce package -> vscode:prepublish ->
webpack -> npm run gen), never a bare `npm install` inside server/.

localInfoPrefix is a cross-cutting completion/hover/symbol convention,
not hover-specific -- moved it to utils/variables.ts (alongside the
other small shared constants) instead of having completion.ts and
workspaceSymbol.ts import it from an unrelated provider file.
