+++ server/package.json
@@ -7,8 +7,7 @@
     "node": "*"
   },
   "scripts": {
-    "gen": "bash scripts/build-ascot-component.sh",
-    "prepare": "npm run gen"
+    "gen": "bash scripts/build-ascot-component.sh"
   },
   "dependencies": {
     "node-html-parser": "^7.1.0",
